split Game into Game and Level
