+++ untitled-sdl-tank-game/Game.h
@@ -1,18 +1,8 @@
 #pragma once
 
-#include "GameObject.h"
-#include "Player.h"
-#include "Enemy.h"
-#include "Bullet.h"
-#include "Flame.h"
-#include "Bomb.h"
+#include "Level.h"
+#include "TextureDataStruct.h"
 #include "menu.h"
-#include "HUD.h"
-#include "MyTexture.h"
-#include "MyText.h"
-#include "PatrollingEnemy.h"
-#include "Crate.h"
-#include "game_settings.h"
 
 #include<SDL.h>
 #include<SDL_image.h>
@@ -21,86 +11,26 @@
 
 #include<vector>
 
-struct TextureDataStruct
-{
-    MyTexture terrainTex;
-    MyTexture playerTexture;
-    MyTexture enemyTexture;
-    MyTexture brickTexture;
-    MyTexture bulletTexture;
-    MyTexture flameTexture;
-    MyTexture helpScreenTexture;
-    MyTexture bombTexture;
-    MyTexture bombInfo;
-    MyTexture cannonInfo;
-    MyTexture someText;
-    MyTexture treeTexture;
-    MyTexture treeTexture2;
-    MyTexture coinTexture;
-    MyTexture sparkTexture;
-    MyTexture grassTexture;
-    MyTexture crateTexture;
-    MyTexture lakeTexture;
-    MyTexture woodenHouseTexture;
-    MyTexture machineGunBulletTexture;
-};
-
 class Game
 {
-private:
-    void initGrass();
-
 public:
     // SDL fields
     SDL_Window* window;
     SDL_Renderer* renderer;
     SDL_Surface* screenSurface;
-    TTF_Font* ttfFont;
-
-    // music
     Mix_Music *music;
+    TTF_Font* ttfFont;
 
     // GUI (menu etc.)
     GameMenu gameMenu;
-    HUD hud;    
-
-    // text
-    MyText* someText;
-
-    // game objects
-    GameObject terrain;
-    Player player;	
-
-    // copyable objects
-    Bullet bulletTemplate;
-    Flame flameTemplate;
-    Bomb bombTemplate;
-    GameObject coinTemplate;
-    GameObject sparkTemplate;
-    GameObject grassTemplate;
-    Crate crateTemplate;
-    GameObject lake;
-    GameObject woodenHouse;
-    Bullet machineGunBulletTemplate;
-
-    // groups of objects
-    std::vector<Enemy> enemies;
-	std::vector<PatrollingEnemy> patrollingEnemies;
-    std::vector<Bullet> bullets;
-    std::vector<Flame> flames;
-    std::vector<GameObject> bricks;
-    std::vector<Bomb> bombs;
-    std::vector<GameObject> trees;
-    std::vector<GameObject> coins;
-    std::vector<GameObject> sparks;
-    std::vector<GameObject> grassTiles;
-    std::vector<Crate> crates;
+    
+    // Level
+    Level* lev;
     
     // texture struct	
     TextureDataStruct texDataStruct;
 
     // other
-
     int screenWidth;
     int screenHeight;
 
@@ -109,27 +39,20 @@ public:
 
     unsigned long int mainLoopCnt = 0;
 
-    // constructors and destructors
-
     Game();
     ~Game();
-
-    // public methods
 
     bool initSDL();
 
     void loadTextures();
     void createGameMenu();
     bool initGame();
-    void mainLoop();
-    bool runGame();	
-    bool endGame();
-    void titleScreenLoop();
+    bool runGame();
+    void setLevel(Level* l);
 
-    void printText(std::string text, int x, int y);
+    bool endGame();
     
     // utils
-
     /**
      * get X coord in screen axis system
      * \param localPosX - x coord in local axis system
@@ -141,6 +64,4 @@ public:
      */
     int getPosYOnScreen(float localPosY);
 };
-
-void addFlame(const Flame& f, std::vector<Flame>& flames, float x, float y);
