+++ untitled-sdl-tank-game/PatrollingEnemy.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "Enemy.h"
 #include "utils/utils.h"
