+++ untitled-sdl-tank-game/TextureDataStruct.h
@@ -1,0 +1,27 @@
+#pragma once
+
+#include "MyTexture.h"
+
+struct TextureDataStruct
+{
+    MyTexture terrainTex;
+    MyTexture playerTexture;
+    MyTexture enemyTexture;
+    MyTexture brickTexture;
+    MyTexture bulletTexture;
+    MyTexture flameTexture;
+    MyTexture helpScreenTexture;
+    MyTexture bombTexture;
+    MyTexture bombInfo;
+    MyTexture cannonInfo;
+    MyTexture someText;
+    MyTexture treeTexture;
+    MyTexture treeTexture2;
+    MyTexture coinTexture;
+    MyTexture sparkTexture;
+    MyTexture grassTexture;
+    MyTexture crateTexture;
+    MyTexture lakeTexture;
+    MyTexture woodenHouseTexture;
+    MyTexture machineGunBulletTexture;
+};
