+++ untitled-sdl-tank-game/main.cpp
@@ -5,8 +5,10 @@
 
 int main(int argc,char* args[])
 {
-    Game g;
-    g.runGame();
-    g.endGame();
+    Game game;
+    Level level(game);
+    game.setLevel(&level);
+    game.runGame();
+    game.endGame();
     return 0;
 }
